refactor: comment-delete feature simplify (#337)
